refactor : ticket - Service 에서 Event 영역 분리 #210
