+++ ticket-service/src/main/java/com/boeingmerryho/business/ticketservice/infrastructure/adapter/kafka/PaymentEventConsumer.java
@@ -3,7 +3,7 @@ package com.boeingmerryho.business.ticketservice.infrastructure.adapter.kafka;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
-import com.boeingmerryho.business.ticketservice.application.user.TicketService;
+import com.boeingmerryho.business.ticketservice.application.user.TicketEventService;
 import com.boeingmerryho.business.ticketservice.infrastructure.adapter.kafka.dto.response.PaymentListenerDto;
 
 import lombok.RequiredArgsConstructor;
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class PaymentEventConsumer {
 
-	private final TicketService ticketService;
+	private final TicketEventService ticketEventService;
 
 	@KafkaListener(
 		topics = "ticket-payment-process",
@@ -20,6 +20,6 @@ public class PaymentEventConsumer {
 		containerFactory = "paymentKafkaListenerContainerFactory"
 	)
 	public void consume(PaymentListenerDto dto) {
-		ticketService.handlePaymentEvent(dto);
+		ticketEventService.handlePaymentEvent(dto);
 	}
 }
