+++ ticket-service/src/main/java/com/boeingmerryho/business/ticketservice/infrastructure/adapter/kafka/SeatEventConsumer.java
@@ -3,10 +3,9 @@ package com.boeingmerryho.business.ticketservice.infrastructure.adapter.kafka;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
-import com.boeingmerryho.business.ticketservice.application.user.TicketService;
+import com.boeingmerryho.business.ticketservice.application.user.TicketEventService;
 import com.boeingmerryho.business.ticketservice.infrastructure.adapter.kafka.dto.response.SeatListenerDto;
 import com.boeingmerryho.business.ticketservice.infrastructure.auditing.CustomAuditorAware;
-import com.fasterxml.jackson.core.JsonProcessingException;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,7 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class SeatEventConsumer {
 
-	private final TicketService ticketService;
+	private final TicketEventService ticketEventService;
 
 	@KafkaListener(
 		topics = "ticket-created",
@@ -27,7 +26,7 @@ public class SeatEventConsumer {
 		try {
 			Long userId = Long.parseLong(dto.seatsInfo().get(0).userId());
 			CustomAuditorAware.setAuditor(userId);
-			ticketService.handleSeatEvent(dto);
+			ticketEventService.handleSeatEvent(dto);
 		} finally {
 			CustomAuditorAware.clear();
 		}
